pyproject.toml
[project]
name = "storm-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lyzr-agents[storm]>=0.1.2",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
]
